pyproject.toml
[project]
name = "aurora"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.4",
    "openai>=2.7.2",
    "requests>=2.32.5",
    "ruff>=0.14.4",
    "python-dotenv>=1.0.0",
]
